chore: move simplification into submodule
