+++ compiler/noirc_evaluator/src/ssa/opt/loop_invariant.rs
@@ -83,14 +83,10 @@ use crate::ssa::{
     ir::{
         basic_block::BasicBlockId,
         cfg::ControlFlowGraph,
-        dfg::simplify::SimplifyResult,
         dom::DominatorTree,
         function::Function,
         function_inserter::FunctionInserter,
-        instruction::{
-            Binary, BinaryOp, ConstrainError, Instruction, InstructionId,
-            binary::{BinaryEvaluationResult, eval_constant_binary_op},
-        },
+        instruction::{Instruction, InstructionId},
         integer::IntegerConstant,
         post_order::PostOrder,
         types::{NumericType, Type},
@@ -100,9 +96,10 @@ use crate::ssa::{
 };
 use acvm::{FieldElement, acir::AcirField};
 use fxhash::{FxHashMap as HashMap, FxHashSet as HashSet};
-use noirc_errors::call_stack::CallStackId;
 
 use super::unrolling::{Loop, Loops};
+
+mod simplify;
 
 impl Ssa {
     /// See [`loop_invariant`][self] module for more information.
@@ -580,6 +577,10 @@ impl<'f> LoopInvariantContext<'f> {
             || self.can_be_hoisted_from_loop_bounds(&instruction)
     }
 
+    fn is_loop_invariant(&self, value_id: &ValueId) -> bool {
+        !self.defined_in_loop.contains(value_id) || self.loop_invariants.contains(value_id)
+    }
+
     /// Check [can_be_hoisted] with extra control dependence information that
     /// lives within the context of this pass.
     fn can_be_hoisted_with_control_dependence(
@@ -690,369 +691,6 @@ impl<'f> LoopInvariantContext<'f> {
     /// that simplification does not require that current block is pure to be simplified.
     fn can_simplify_control_dependent_instruction(&self) -> bool {
         !self.current_block_control_dependent && self.current_block_executes
-    }
-
-    /// Some instructions can take advantage of that our induction variable has a fixed minimum/maximum,
-    /// For instance operations can be transformed from a checked operation to an unchecked operation.
-    ///
-    /// Checked operations require more bytecode and thus we aim to minimize their usage wherever possible.
-    ///
-    /// For example, if one side of an add/mul operation is a constant and the other is an induction variable
-    /// with a known upper bound, we know whether that binary operation will ever overflow.
-    /// If we determine that an overflow is not possible we can convert the checked operation to unchecked.
-    ///
-    /// The function returns false if the instruction must be added to the block
-    fn simplify_from_loop_bounds(&mut self, instruction_id: InstructionId, header: bool) -> bool {
-        // Simplify the instruction and update it in the DFG.
-        match self.simplify_induction_variable(instruction_id, header) {
-            SimplifyResult::SimplifiedTo(id) => {
-                let results =
-                    self.inserter.function.dfg.instruction_results(instruction_id).to_vec();
-                assert!(results.len() == 1);
-                self.inserter.map_value(results[0], id);
-                true
-            }
-            SimplifyResult::SimplifiedToInstruction(instruction) => {
-                self.inserter.function.dfg[instruction_id] = instruction;
-                false
-            }
-            SimplifyResult::Remove => true,
-            SimplifyResult::None => false,
-            _ => unreachable!(
-                "ICE - loop bounds simplification should only simplify to a value or an instruction"
-            ),
-        }
-    }
-
-    /// Simplify 'assert(lhs < rhs)' into 'assert(max(lhs) < rhs)' if lhs is an induction variable and rhs a loop invariant
-    /// For this simplification to be valid, we need to ensure that the induction variable takes all the values from min(induction) up to max(induction)
-    /// This means that the assert must be executed  at each loop iteration, and that the loop processes all the iteration space
-    /// This is ensured via control dependence and the check for break patterns, before calling this function.
-    fn simplify_induction_in_constrain(
-        &mut self,
-        lhs: ValueId,
-        rhs: ValueId,
-        err: &Option<ConstrainError>,
-        call_stack: CallStackId,
-    ) -> SimplifyResult {
-        let mut extract_variables = |rhs, lhs| {
-            let rhs_true =
-                self.inserter.function.dfg.get_numeric_constant(rhs)? == FieldElement::one();
-            let (is_left, min, max, binary) = self.extract_induction_and_invariant(lhs)?;
-            match (is_left, rhs_true) {
-                (true, true) => Some((max, binary.rhs)),
-                (false, true) => Some((binary.lhs, min)),
-                _ => None,
-            }
-        };
-        let Some((new_lhs, new_rhs)) = extract_variables(rhs, lhs) else {
-            return SimplifyResult::None;
-        };
-        let new_binary =
-            Instruction::Binary(Binary { lhs: new_lhs, rhs: new_rhs, operator: BinaryOp::Lt });
-        // The new comparison can be safely hoisted to the pre-header because it is loop invariant and control independent
-        let comparison_results = self
-            .inserter
-            .function
-            .dfg
-            .insert_instruction_and_results(new_binary, self.pre_header(), None, call_stack)
-            .results();
-        assert!(comparison_results.len() == 1);
-        SimplifyResult::SimplifiedToInstruction(Instruction::Constrain(
-            comparison_results[0],
-            rhs,
-            err.clone(),
-        ))
-    }
-
-    /// Replace 'assert(invariant != induction)' with assert((invariant < min(induction) || (invariant > max(induction)))
-    /// For this simplification to be valid, we need to ensure that the induction variable takes all the values from min(induction) up to max(induction)
-    /// This means that the assert must be executed  at each loop iteration, and that the loop processes all the iteration space
-    /// This is ensured via control dependence and the check for break patterns, before calling this function.
-    fn simplify_not_equal_constraint(
-        &mut self,
-        lhs: &ValueId,
-        rhs: &ValueId,
-        err: &Option<ConstrainError>,
-        call_stack: CallStackId,
-    ) -> SimplifyResult {
-        let (invariant, upper, lower) = match self.match_induction_and_invariant(lhs, rhs) {
-            Some((true, upper, lower)) => (rhs, upper, lower),
-            Some((false, upper, lower)) => (lhs, upper, lower),
-            _ => return SimplifyResult::None,
-        };
-
-        let mut insert_binary_to_preheader = |lhs, rhs, operator| {
-            let binary = Instruction::Binary(Binary { lhs, rhs, operator });
-            let results = self
-                .inserter
-                .function
-                .dfg
-                .insert_instruction_and_results(binary, self.pre_header(), None, call_stack)
-                .results();
-            assert!(results.len() == 1);
-            results[0]
-        };
-        // The comparisons can be safely hoisted to the pre-header because they are loop invariant and control independent
-        let check_lower_bound = insert_binary_to_preheader(*invariant, lower, BinaryOp::Lt);
-        let check_upper_bound = insert_binary_to_preheader(upper, *invariant, BinaryOp::Lt);
-        let check_bounds =
-            insert_binary_to_preheader(check_lower_bound, check_upper_bound, BinaryOp::Or);
-
-        SimplifyResult::SimplifiedToInstruction(Instruction::Constrain(
-            check_bounds,
-            self.true_value,
-            err.clone(),
-        ))
-    }
-
-    /// Returns the binary instruction only if the input value refers to a binary instruction with invariant and induction variables as operands
-    /// The return values are:
-    /// - a boolean indicating if the induction variable is on the lhs
-    /// - the minimum and maximum values of the induction variable, coming from the loop bounds
-    /// - the binary instruction itself
-    fn extract_induction_and_invariant(
-        &mut self,
-        value: ValueId,
-    ) -> Option<(bool, ValueId, ValueId, Binary)> {
-        let Value::Instruction { instruction, .. } = &self.inserter.function.dfg[value] else {
-            return None;
-        };
-        let Instruction::Binary(binary) = self.inserter.function.dfg[*instruction].clone() else {
-            return None;
-        };
-        self.match_induction_and_invariant(&binary.lhs, &binary.rhs)
-            .map(|(is_left, min, max)| (is_left, min, max, binary))
-    }
-
-    /// If the inputs are an induction and loop invariant variables, it returns
-    /// the maximum and minimum values of the induction variable, based on the loop bounds,
-    /// and a boolean indicating if the induction variable is on the lhs or rhs (true for lhs)
-    fn match_induction_and_invariant(
-        &mut self,
-        lhs: &ValueId,
-        rhs: &ValueId,
-    ) -> Option<(bool, ValueId, ValueId)> {
-        let (is_left, lower, upper) = match (
-            self.current_induction_variables.get(lhs),
-            self.current_induction_variables.get(rhs),
-        ) {
-            (_, Some((lower, upper))) => Some((false, lower, upper)),
-            (Some((lower, upper)), _) => Some((true, lower, upper)),
-            _ => None,
-        }?;
-
-        assert!(self.current_block_executes, "executing a non executable loop");
-
-        let (upper_field, upper_type) = upper.dec().into_numeric_constant();
-        let (lower_field, lower_type) = lower.into_numeric_constant();
-
-        let min_iter = self.inserter.function.dfg.make_constant(lower_field, lower_type);
-        let max_iter = self.inserter.function.dfg.make_constant(upper_field, upper_type);
-        if (is_left && self.is_loop_invariant(rhs)) || (!is_left && self.is_loop_invariant(lhs)) {
-            return Some((is_left, min_iter, max_iter));
-        }
-        None
-    }
-
-    /// Simplify certain instructions using the lower/upper bounds of induction variables
-    fn simplify_induction_variable(
-        &mut self,
-        instruction_id: InstructionId,
-        header: bool,
-    ) -> SimplifyResult {
-        let (instruction, call_stack) = self.inserter.map_instruction(instruction_id);
-        match &instruction {
-            Instruction::Binary(binary) => {
-                self.simplify_induction_variable_in_binary(binary, header)
-            }
-            Instruction::Constrain(x, y, err) => {
-                // Ensure the loop is fully executed
-                if self.no_break && self.can_simplify_control_dependent_instruction() {
-                    self.simplify_induction_in_constrain(*x, *y, err, call_stack)
-                } else {
-                    SimplifyResult::None
-                }
-            }
-            Instruction::ConstrainNotEqual(x, y, err) => {
-                // Ensure the loop is fully executed
-                if self.no_break && self.can_simplify_control_dependent_instruction() {
-                    self.simplify_not_equal_constraint(x, y, err, call_stack)
-                } else {
-                    SimplifyResult::None
-                }
-            }
-            _ => SimplifyResult::None,
-        }
-    }
-
-    fn is_loop_invariant(&self, value_id: &ValueId) -> bool {
-        !self.defined_in_loop.contains(value_id) || self.loop_invariants.contains(value_id)
-    }
-
-    /// If the inputs are an induction variable and a constant, it returns
-    /// the constant value, the maximum and minimum values of the induction variable, based on the loop bounds,
-    /// and a boolean indicating if the induction variable is on the lhs or rhs (true for lhs)
-    /// if `only_outer_induction`, we only consider outer induction variables, else we also consider the induction variables from the current loop.
-    fn match_induction_and_constant(
-        &self,
-        lhs: &ValueId,
-        rhs: &ValueId,
-        only_outer_induction: bool,
-    ) -> Option<(bool, IntegerConstant, IntegerConstant, IntegerConstant)> {
-        let lhs_const = self.inserter.function.dfg.get_integer_constant(*lhs);
-        let rhs_const = self.inserter.function.dfg.get_integer_constant(*rhs);
-        match (
-            lhs_const,
-            rhs_const,
-            self.current_induction_variables
-                .get(lhs)
-                .and_then(|v| if only_outer_induction { None } else { Some(v) })
-                .or(self.outer_induction_variables.get(lhs)),
-            self.current_induction_variables
-                .get(rhs)
-                .and_then(|v| if only_outer_induction { None } else { Some(v) })
-                .or(self.outer_induction_variables.get(rhs)),
-        ) {
-            // LHS is a constant, RHS is the induction variable with a known lower and upper bound.
-            (Some(lhs), None, None, Some((lower_bound, upper_bound))) => {
-                Some((false, lhs, *lower_bound, *upper_bound))
-            }
-            // RHS is a constant, LHS is the induction variable with a known lower an upper bound
-            (None, Some(rhs), Some((lower_bound, upper_bound)), None) => {
-                Some((true, rhs, *lower_bound, *upper_bound))
-            }
-            _ => None,
-        }
-    }
-
-    /// Given a constant `c` and an induction variable `i`:
-    /// - Replace comparisons `i < c` by true if `max(i) < c`, and false if `min(i) >= c`
-    /// - Replace comparisons `c < i` by true if `min(i) > c`, and false if `max(i) <= c`
-    /// - Replace equalities `i == c` by false if `min(i) > c or max(i) < c`
-    /// - Replace checked operations with unchecked version if the induction variable bounds prove that the operation will not overflow
-    ///
-    /// `header` indicates if we are in the loop header where loop bounds do not apply yet
-    fn simplify_induction_variable_in_binary(
-        &mut self,
-        binary: &Binary,
-        header: bool,
-    ) -> SimplifyResult {
-        // Checks the operands are an induction variable and a constant
-        // Note that here we allow all_induction_variables
-        let operand_type = self.inserter.function.dfg.type_of_value(binary.lhs).unwrap_numeric();
-
-        let Some((is_induction_var_lhs, value, lower_bound, upper_bound)) =
-            self.match_induction_and_constant(&binary.lhs, &binary.rhs, header)
-        else {
-            return SimplifyResult::None;
-        };
-
-        // Handle arithmetic operations
-        if let Some((lhs, rhs)) = match binary.operator {
-            BinaryOp::Add { unchecked }
-            | BinaryOp::Sub { unchecked }
-            | BinaryOp::Mul { unchecked }
-                if unchecked =>
-            {
-                return SimplifyResult::None;
-            }
-            BinaryOp::Sub { .. } => {
-                if is_induction_var_lhs {
-                    Some((lower_bound, value))
-                } else {
-                    Some((value, upper_bound))
-                }
-            }
-            BinaryOp::Add { .. } | BinaryOp::Mul { .. } => Some((value, upper_bound)),
-            BinaryOp::Div | BinaryOp::Mod => return SimplifyResult::None,
-            _ => None,
-        } {
-            // We evaluate this expression using the upper bounds (or lower in the case of sub)
-            // of its inputs to check whether it will ever overflow.
-            // If `eval_constant_binary_op` won't overflow we can simplify the instruction to an unchecked version.
-            let lhs = lhs.into_numeric_constant().0;
-            let rhs = rhs.into_numeric_constant().0;
-            match eval_constant_binary_op(lhs, rhs, binary.operator, operand_type) {
-                BinaryEvaluationResult::Success(..) => {
-                    // Unchecked version of the binary operation
-                    let unchecked = Instruction::Binary(Binary {
-                        operator: binary.operator.into_unchecked(),
-                        lhs: binary.lhs,
-                        rhs: binary.rhs,
-                    });
-                    return SimplifyResult::SimplifiedToInstruction(unchecked);
-                }
-                BinaryEvaluationResult::CouldNotEvaluate | BinaryEvaluationResult::Failure(..) => {
-                    return SimplifyResult::None;
-                }
-            }
-        }
-
-        // Handle comparisons
-        match binary.operator {
-            BinaryOp::Eq => {
-                if value >= upper_bound || value < lower_bound {
-                    SimplifyResult::SimplifiedTo(self.false_value)
-                } else {
-                    SimplifyResult::None
-                }
-            }
-            BinaryOp::Lt => match is_induction_var_lhs {
-                true if upper_bound <= value => SimplifyResult::SimplifiedTo(self.true_value),
-                true if lower_bound >= value => SimplifyResult::SimplifiedTo(self.false_value),
-                false if lower_bound > value => SimplifyResult::SimplifiedTo(self.true_value),
-                false if upper_bound <= value.inc() => {
-                    SimplifyResult::SimplifiedTo(self.false_value)
-                }
-                _ => SimplifyResult::None,
-            },
-            _ => SimplifyResult::None,
-        }
-    }
-
-    /// Checks whether a binary operation can be evaluated using the bounds of a given loop induction variables.
-    ///
-    /// If it cannot be evaluated, it means that we either have a dynamic loop bound or
-    /// that the operation can potentially overflow during a given loop iteration.
-    fn can_evaluate_binary_op(&self, binary: &Binary) -> bool {
-        match binary.operator {
-            // An unchecked operation cannot overflow, so it can be safely evaluated
-            BinaryOp::Add { unchecked: true }
-            | BinaryOp::Mul { unchecked: true }
-            | BinaryOp::Sub { unchecked: true } => true,
-            BinaryOp::Div | BinaryOp::Mod => {
-                // Division can be evaluated if we ensure that the divisor cannot be zero
-                let Some((left, value, lower, upper)) =
-                    self.match_induction_and_constant(&binary.lhs, &binary.rhs, true)
-                else {
-                    // Not a constant vs non-constant case, we cannot evaluate it.
-                    return false;
-                };
-
-                if left {
-                    // If the induction variable is on the LHS, we're dividing with a constant.
-                    if !value.is_zero() {
-                        return true;
-                    }
-                } else {
-                    // Otherwise we are dividing a constant with the induction variable, and we have to check whether
-                    // at any point in the loop the induction variable can be zero.
-                    let can_be_zero = lower.is_negative() && !upper.is_negative()
-                        || lower.is_zero()
-                        || upper.is_zero();
-
-                    if !can_be_zero {
-                        return true;
-                    }
-                }
-
-                false
-            }
-            // Some checked operations can be safely evaluated, depending on the loop bounds, but in that case,
-            // they would have been already converted to unchecked operation in `simplify_induction_variable_in_binary()`
-            _ => false,
-        }
     }
 
     /// Loop invariant hoisting only operates over loop instructions.
